Add install verification option for installer download
Description
===========

Some Unity installer metafiles (.ini) contain incorrect checksums for
the installer. Or maybe the user wants to skip the step of verification
to shave of a few seconds of the install process. This patch adds an
option `--[no]-verify` to the `uvm install` command.

The download logic is no longer just a function. This patch adds a new
`Loader` struct in `uvm_core::install::installer::loader`. This structs
sets an internal `bool` `verify` to `true` by default. This value can be
overridden if wished.

Changes
=======

![ADD] `-[-no]-verify` commandline option for `uvm install` command
![IMPROVE] internal installer download logic
