+++ uvm_core/src/install/installer/mod.rs
@@ -5,6 +5,8 @@ mod other;
 #[cfg(target_os = "windows")]
 mod windows;
 
+mod loader;
+
 #[cfg(target_os = "macos")]
 use self::macos as sys;
 #[cfg(not(any(target_os = "windows", target_os = "macos")))]
@@ -14,3 +16,4 @@ use self::windows as sys;
 
 pub use self::sys::install_editor;
 pub use self::sys::install_module;
+pub use self::loader::Loader;
